update: styles and info in dancer and representative profiles
