+++ styles/dashboard.module.css
@@ -163,6 +163,7 @@
   transition: transform 0.5s;
   margin: auto;
   background-color: black;
+  margin-bottom: 12px;
 }
 
 .button:hover{
@@ -310,7 +311,7 @@
   font-size: 20px;
   color: white;
   width: 300px;
-  height: 300px;
+  height: auto;
   background-color:black;
   padding: 16px;
   margin-top: 30px;
@@ -337,7 +338,7 @@
 @media screen and (max-width: 460px){
   .danceCard {
     width: 260px;
-    height: 180px;
+    height: auto;
   }
   .danceCardNoPay {
     width: 260px;
@@ -609,7 +610,8 @@
 .dancersCont{
   display: flex;
   flex-direction: column;
-  min-height: 700px;
+  justify-content: flex-start;
+  align-items: center;
   height: auto;
   gap: 60px;
 }
@@ -619,3 +621,12 @@
   flex-direction:column;
   gap: 30px;
 }
+
+.details{
+  margin-left: 1.5rem;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  gap: 1rem;
+  color: white;
+}
+
